+++ src/runtime_src/xdp/profile/plugin/device_offload/device_offload_plugin.cpp
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2020 Xilinx, Inc
+ * Copyright (C) 2020-2021 Xilinx, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may
  * not use this file except in compliance with the License. A copy of the
@@ -210,7 +210,8 @@ namespace xdp {
 
     // If trace is enabled, set up trace.  Otherwise just keep the offloader
     //  for reading the counters.
-    if (xrt_core::config::get_data_transfer_trace() != "off") {
+    if (xrt_core::config::get_data_transfer_trace() != "off" ||
+        xrt_core::config::get_device_trace() != "off") {
       bool init_successful =
         offloader->read_trace_init(m_enable_circular_buffer) ;
 
@@ -283,6 +284,9 @@ namespace xdp {
     // Collect all the profiling options from xrt.ini
     std::string data_transfer_trace = 
       xrt_core::config::get_data_transfer_trace() ;
+    if (data_transfer_trace == "off") {
+      data_transfer_trace = xrt_core::config::get_device_trace() ;
+    }
     std::string stall_trace = xrt_core::config::get_stall_trace() ;
 
     // Set up the hardware trace option